# openapi.yaml
openapi: 3.0.3
info:
  title: P0rtal RDP Broker API
  description: API for managing RDP sessions through the P0rtal broker proxy.
  version: 3.0.0
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        API key passed as Bearer token. Used for management API endpoints
        (access control, public keys, recordings, secrets).
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Signed JWT for user access API (POST /api/sessions). The JWT must
        contain a `sub` claim (user email) and be signed with a private key
        whose public key is registered via /api/public-keys.
  schemas:
...
    HostConfig:
      type: object
      required: [hostname, ip]
      properties:
        hostname:
          type: string
          example: mike-rdp
        ip:
          type: string
          example: 10.1.0.7
        port:
          type: integer
          default: 3389
          example: 3389
        domain:
          type: string
          example: P0-LAB-1
    UserConfig:
      type: object
      required: [email, username, secret]
      properties:
        email:
          type: string
          format: email
          example: miguel.campos@p0.dev
        username:
          type: string
          description: Windows username for RDP login
          example: rdpadmin
        secret:
          type: string
          description: Secret name in Google Secret Manager
          example: p0_windows_domain_rdpadmin
    Grant:
      type: object
      required: [host, user]
      properties:
        host:
          $ref: "#/components/schemas/HostConfig"
        user:
          $ref: "#/components/schemas/UserConfig"
    RevokeRequest:
      type: object
      required: [email, hostname]
      properties:
        email:
          type: string
          format: email
          example: miguel.campos@p0.dev
        hostname:
          type: string
          example: mike-rdp
    PublicKeyRequest:
      type: object
      required: [email, public_key]
      properties:
        email:
          type: string
          format: email
          example: miguel.campos@p0.dev
        public_key:
          type: string
          description: Public key in PEM, SSH authorized key, or base64 DER format
          example: "ssh-rsa AAAA..."
    PublicKeyDeleteRequest:
      type: object
      required: [email, fingerprint]
      properties:
        email:
          type: string
          format: email
        fingerprint:
          type: string
          description: SHA256 fingerprint of the key to delete
          example: "SHA256:abc..."
    PublicKeyEntry:
      type: object
      properties:
        email:
          type: string
          format: email
        public_key:
          type: string
        fingerprint:
          type: string
...
paths:
...
  /api/access:
    get:
      summary: List all access grants
      description: Returns all grants with full host + user configuration.
      tags: [Access Control]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grant"
  /api/grant:
    post:
      summary: Grant access
      description: >-
        Creates or updates an access grant with full host + user configuration.
        Idempotent — re-granting the same (email, hostname) updates the config.
      tags: [Access Control]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Grant"
      responses:
        "200":
          description: Access granted
        "400":
          description: Missing required fields
  /api/revoke:
    post:
      summary: Revoke access
      description: >-
        Revokes access for (email, hostname) and terminates active sessions.
      tags: [Access Control]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
      responses:
        "200":
          description: Access revoked
        "404":
          description: Grant not found
  /api/public-keys:
    get:
      summary: List all public keys
      tags: [Public Keys]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of registered public keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicKeyEntry"
    post:
      summary: Register a public key
      description: >-
        Adds a public key for a user email. A user can have multiple keys.
        Idempotent — adding the same key is a no-op.
      tags: [Public Keys]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicKeyRequest"
      responses:
        "200":
          description: Key registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  fingerprint:
                    type: string
        "400":
          description: Invalid email or public key
    delete:
      summary: Remove a public key
      description: Removes a specific key by email + fingerprint.
      tags: [Public Keys]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicKeyDeleteRequest"
      responses:
        "200":
          description: Key removed
        "404":
          description: Key not found
...